# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TS Screen Legacy API",
        "description": "Compatible with 3 Flutter apps. Swagger grouped by app: Customer (Phone), Projector (TV), Admin. Shared routes appear in every app that calls them. App receives JSON string (text/html). Swagger Try it out receives application/json. status=1 success, -2 show msg, -1 OTP error.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "paths": {
...
        "/home/CreatePacket": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin tạo gói (thêm is_trial, giá 1/6/12 tháng, limit_qty, limit_capacity bytes)",
                "operationId": "56e27aa7f0dca73d4f103cd0ab41066e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name_packet": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "price_6_month": {
                                        "type": "string"
                                    },
                                    "price_12_month": {
                                        "type": "string"
                                    },
                                    "month_qty": {
                                        "type": "string"
                                    },
                                    "day_qty": {
                                        "type": "string"
                                    },
                                    "year_qty": {
                                        "type": "string"
                                    },
                                    "picture": {
                                        "type": "string"
                                    },
                                    "detail": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "limit_qty": {
                                        "description": "Số TV tối đa",
                                        "type": "string"
                                    },
                                    "limit_capacity": {
                                        "description": "Bytes; 1–1024 = GB",
                                        "type": "string"
                                    },
                                    "account_id": {
                                        "type": "string"
                                    },
                                    "is_trial": {
                                        "type": "string"
                                    },
                                    "is_business": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{status:1}"
                    }
                }
            }
        },
        "/home/UpdatePacket_ById/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin sửa gói",
                "operationId": "03105a2e620e882a082fd9983dbffa5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name_packet": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "price_6_month": {
                                        "type": "string"
                                    },
                                    "price_12_month": {
                                        "type": "string"
                                    },
                                    "limit_qty": {
                                        "type": "string"
                                    },
                                    "limit_capacity": {
                                        "type": "string"
                                    },
                                    "is_trial": {
                                        "type": "string"
                                    },
                                    "is_business": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{status:1}"
                    }
                }
            }
        },
        "/home/DeletePacket_ById/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin xóa gói (soft delete)",
                "operationId": "3b5948dd11b1477f159e0471a93bf566",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:1}"
                    }
                }
            }
        },
...
    },
...
}